Chrome 53 is removing DH
If DH is present in the cipher list and any previous cipher in the list
doesn't match with the chrome internal cipher list, the connection will
result in a ERR_SSL_OBSOLETE_CIPHER connection error. The only fix we
found has been to totally remove DHE from the cipherlist.

Reference:
  https://codereview.chromium.org/2056343006
  https://www.chromestatus.com/features/5128908798164992
  https://developers.google.com/web/updates/2016/08/chrome-53-deprecations?hl=en#dhe-based-ciphers-being-phased-out

Signed-off-by: Arthur Gautier <baloo@gandi.net>